+++ features/handler/product.go
@@ -1,11 +1,15 @@
 package handler
 
 import (
+	"io"
 	"lokasani/entity/request"
 	"lokasani/entity/response"
 	"lokasani/features/services"
+	consts "lokasani/helpers/const"
+	"lokasani/helpers/middleware"
+	"strconv"
 
-	"github.com/labstack/echo"
+	"github.com/labstack/echo/v4"
 )
 
 type ProductHandler struct {
@@ -17,8 +21,30 @@ func NewProductHandler(iProductService services.IProductService) *ProductHandler
 }
 
 func (pr *ProductHandler) CreateProduct(c echo.Context) error {
+	_, role, _, err := middleware.ExtractToken(c)
+    if err != nil {
+        return response.NewErrorResponse(c, err)
+    }
+    if role != consts.ProductCreator {
+		return response.NewErrorResponse(c, echo.ErrUnauthorized)
+	}
+	// if roleId != 1 {
+	// return response.NewErrorResponse(c, echo.ErrUnauthorized)
+	// }
+
 	var input request.Product
 	c.Bind(&input)
+	file, err := c.FormFile("file")
+	if file != nil {
+		src, err := file.Open()
+		if err != nil {
+			return response.NewErrorResponse(c, err)
+		}
+		defer src.Close()
+
+		fileBytes, err := io.ReadAll(src)
+		input.File = fileBytes
+	}
 	res, err := pr.productService.CreateProduct(&input)
 	if err != nil {
 		return response.NewErrorResponse(c, err)
@@ -28,12 +54,72 @@ func (pr *ProductHandler) CreateProduct(c echo.Context) error {
 
 func (pr *ProductHandler) GetAllProduct(c echo.Context) error {
 	nameFilter := c.QueryParam("name")
+	page, _ := strconv.Atoi(c.QueryParam("page"))
+    pageSize, _ := strconv.Atoi(c.QueryParam("pageSize"))
 
-    res, err := pr.productService.GetAllProduct(nameFilter)
-    if err != nil {
-        return response.NewErrorResponse(c, err)
+    if page <= 0 {
+        page = 1
     }
-    return response.NewSuccessResponse(c, res)
+
+    if pageSize <= 0 {
+        pageSize = 10
+    }
+
+	res, totalItems, err := pr.productService.GetAllProduct(nameFilter, page, pageSize)
+	if err != nil {
+		return response.NewErrorResponse(c, err)
+	}
+
+	currentPage, allPages := pr.productService.CalculatePaginationValues(page, pageSize, totalItems)
+	nextPage := pr.productService.GetNextPage(currentPage, allPages)
+	prevPage := pr.productService.GetPrevPage(currentPage)
+
+	responseData := map[string]interface{}{
+		"allProduct": res,
+		"pagination": map[string]int{
+			"currentPage": currentPage,
+			"nextPage":    nextPage,
+			"prevPage":    prevPage,
+			"allPages":    allPages,
+		},
+	}
+
+	return response.NewSuccessResponse(c, responseData)
+}
+
+func (pr *ProductHandler) GetTrendingProduct(c echo.Context) error {
+	nameFilter := c.QueryParam("name")
+	page, _ := strconv.Atoi(c.QueryParam("page"))
+    pageSize, _ := strconv.Atoi(c.QueryParam("pageSize"))
+
+    if page <= 0 {
+        page = 1
+    }
+
+    if pageSize <= 0 {
+        pageSize = 10
+    }
+
+	res, totalItems, err := pr.productService.GetTrendingProduct(nameFilter, page, pageSize)
+	if err != nil {
+		return response.NewErrorResponse(c, err)
+	}
+
+	currentPage, allPages := pr.productService.CalculatePaginationValues(page, pageSize, totalItems)
+	nextPage := pr.productService.GetNextPage(currentPage, allPages)
+	prevPage := pr.productService.GetPrevPage(currentPage)
+
+	responseData := map[string]interface{}{
+		"allProduct": res,
+		"pagination": map[string]int{
+			"currentPage": currentPage,
+			"nextPage":    nextPage,
+			"prevPage":    prevPage,
+			"allPages":    allPages,
+		},
+	}
+
+	return response.NewSuccessResponse(c, responseData)
 }
 
 func (pr *ProductHandler) GetProduct(c echo.Context) error {
@@ -46,6 +132,14 @@ func (pr *ProductHandler) GetProduct(c echo.Context) error {
 }
 
 func (pr *ProductHandler) UpdateProduct(c echo.Context) error {
+	_, role, _, err := middleware.ExtractToken(c)
+    if err != nil {
+        return response.NewErrorResponse(c, err)
+    }
+    if role != consts.ProductCreator {
+		return response.NewErrorResponse(c, echo.ErrUnauthorized)
+	}
+
 	id := c.Param("id")
 	var input request.Product
 	c.Bind(&input)
@@ -58,6 +152,14 @@ func (pr *ProductHandler) UpdateProduct(c echo.Context) error {
 }
 
 func (pr *ProductHandler) DeleteProduct(c echo.Context) error {
+	_, role, _, err := middleware.ExtractToken(c)
+    if err != nil {
+        return response.NewErrorResponse(c, err)
+    }
+    if role != consts.ProductCreator {
+		return response.NewErrorResponse(c, echo.ErrUnauthorized)
+	}
+
 	id := c.Param("id")
 	res, err := pr.productService.DeleteProduct(id)
 	if err != nil {
